+++ src/components/waitstaff/Activator.tsx
@@ -2,7 +2,7 @@ interface props{
   className?: string;
 }
 
-export default function WaitStaffPanel({ className }: props) {
+export default function Activator({ className }: props) {
   return (
     <main className={`flex-1 p-10 absolute left-1 top-0 ml-96 ${className}`}>
       <button
