--- src/components/waitstaff/App.tsx
@@ -1,12 +1,0 @@
-import WaitStaffPanel from "../../pages/WaitStaffPanel";
-import { OrderProvider } from "./OrderContext";
-
-function App() {
-  return (
-    <OrderProvider>
-      <WaitStaffPanel />
-    </OrderProvider>
-  );
-}
-
-export default App;
