+++ src/components/waitstaff/WaitStaffApp.tsx
@@ -1,0 +1,76 @@
+import Header from "./Header";
+import Activator from "./Activator";
+import { useState } from "react";
+import MaxContainerDishes from "./containers/MaxContainerDishes";
+import MaxContainerTables from "./containers/MaxContainerTables";
+import MaxContainerViewOrders from "./containers/MaxContainerViewOrders";
+import { NotificationPanel } from "./Notifications";
+import { MaxContainerKitchen } from "./containers/MaxContainerKitchen";
+
+export default function WaitStaffApp() {
+  const mesas = Array.from({ length: 16 }, (_, i) => i + 1);
+  const [vistaActual, setVistaActual] = useState<"mesas" | "dishes" | "resumen" | "cocina">("mesas");
+  const [mesaSeleccionada, setMesaSeleccionada] = useState<number>(1);
+
+  return (
+    <>
+      <div className="flex h-screen max-w-[1950px]">
+
+        <NotificationPanel />
+        <Activator />
+
+        <div className="flex flex-col flex-1">
+
+          <Header
+            onTables={() => { setVistaActual("mesas") }}
+            onHeader={() => { setVistaActual("cocina") }}
+            setVistaActual={setVistaActual}
+            vistaActual={vistaActual}
+          />
+
+          <main className="flex-1 overflow-auto p-4 max-w-[1600px]">
+
+            {vistaActual === "mesas" && (
+              <MaxContainerTables
+                mesas={mesas}
+                onIrResumen={(mesa) => {
+                  setMesaSeleccionada(mesa)
+                  setVistaActual("resumen")
+                }}
+                onGoAddDishes={(mesa) => {
+                  setMesaSeleccionada(mesa)
+                  setVistaActual("dishes")
+                }}
+                onSeleccionar={(mesa) => {
+                  setMesaSeleccionada(mesa);
+                  setVistaActual("dishes");
+                }}
+              />
+            )}
+
+            {vistaActual === "dishes" && mesaSeleccionada !== null && (
+              <MaxContainerDishes
+                numeroMesa={mesaSeleccionada}
+                onVolver={() => setVistaActual("mesas")}
+              />
+            )}
+
+            {vistaActual === "resumen" && mesaSeleccionada !== null && (
+              <MaxContainerViewOrders
+                numeroMesa={mesaSeleccionada}
+                onVolver={() => setVistaActual("mesas")}
+                onGoAddDishes={() => setVistaActual("dishes")}
+              />
+            )}
+
+            {vistaActual === "cocina" &&
+
+              <MaxContainerKitchen onVolver={() => { setVistaActual("mesas") }} />
+            }
+
+          </main>
+        </div>
+      </div>
+    </>
+  );
+}
